Update 53-Collect Leaves In A Binary Tree
 Collect Leaves

(https://www.naukri.com/code360/problems/collect-leaves_1281192?leftPanelTabValue=PROBLEM)

Problem statement
You are given a binary tree with ‘N’ nodes. Each node has an integer value associated with it. Your task is to print the leaf nodes and then remove these leaf nodes. Now, print the new leaf nodes and repeat the process until the tree becomes empty.

Note:
You don’t require to remove the leaf nodes. Just print the values in the required order.

The leaf nodes must be printed in a left to right order.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 100
1 <= N <= 10^5
-10^9 <= node data <= 10^9, (where node data != -1).

Time Limit: 1 sec
Sample Input 1:
1
6 -3 3 -1 -1 -1 -1
Sample Output 1:
-3 3
6
Explanation For Sample Output 1:
For the first test case, the given binary tree is depicted below.

In the given tree, the leaf nodes are -3 and 3. When we remove these nodes, we will get 6 as the only node present in the tree. We print it in a separate line as it is the only leaf node left in the tree.
Sample Input 2:
1
1 2 1 -1 -1 -1 -1 
Sample Output 2:
2 1 
1